+++ src/bios/flib/src/x86/mod.rs
@@ -1,2 +1,1 @@
 pub mod flags;
-pub mod interrupts;
